+++ src/include/string.h
@@ -1,7 +1,9 @@
 #ifndef STRING_H
 #define STRING_H
 
-int strlen(const char* s);
+#include <stddef.h>
+
+size_t strlen(const char* s);
 void reverse(char* s);
 
 #endif
